--- components/AlocholType/ATInfo/ATInfo.module.scss
@@ -1,98 +1,0 @@
-@import '@styles/theme.scss';
-
-.container {
-  width: auto;
-  border-radius: 10px;
-  padding: 0.5rem;
-
-  display: flex;
-  flex-direction: row;
-  gap: 5%;
-  @media screen and (min-width: $breakpoint-lg) {
-    // desktop
-    width: 100%;
-    max-height: 40vh;
-    overflow-wrap: break-word;
-  }
-  @media screen and (min-width: $breakpoint-sm) and (max-width: $breakpoint-lg) {
-    // tablet
-    height: 25vh;
-  }
-  @media screen and (max-width: $breakpoint-sm) {
-    //mobile
-    height: auto;
-    flex-direction: column;
-  }
-}
-.alcoholImage {
-  width: 25%;
-  min-width: 200px;
-  height: 320px;
-  object-fit: cover;
-  border-radius: 10px;
-  box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.1);
-}
-.alcoholInfoBox {
-  width: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: left;
-  gap: 0.5rem;
-  overflow-x: hidden;
-  @media screen and (max-width: $breakpoint-sm) {
-    gap: 0.7rem;
-  }
-}
-
-.title {
-  max-width: 100%;
-  margin: 0;
-
-  font-size: $font-size-xl;
-  font-weight: bold;
-
-  @include ellipsisText(2, $font-size-xl);
-
-  @media screen and (max-width: $breakpoint-sm) {
-    margin-top: 0.7rem;
-    font-size: $font-size-xl;
-    @include ellipsisText(1, $font-size-xl);
-  }
-}
-.basic {
-  margin: 0;
-
-  font-size: $font-size-sm;
-  @include ellipsisText(1, $font-size-md);
-
-  @media screen and (max-width: $breakpoint-sm) {
-    font-size: $font-size-sm;
-    @include ellipsisText(1, $font-size-sm);
-  }
-}
-.hashtagBox {
-  max-height: calc((2 * $font-size-xxs) + 30px);
-
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 10px;
-
-  padding-top: 0.2rem;
-
-  font-size: $font-size-xxs;
-  overflow: hidden;
-}
-.hashtag {
-  color: $white;
-  background-color: $gray11;
-  border-radius: 22px;
-  padding: 0.2rem 0.5rem;
-  &:hover {
-    background-color: $accent-color;
-    transition: all 0.2s ease-in-out;
-  }
-}
