--- components/AlocholType/ATInfoDetail/ATInfoDetail.module.scss
@@ -1,21 +1,0 @@
-@import '@styles/theme.scss';
-
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  padding: 0.5rem;
-  width: auto;
-  margin: 1.5rem 0 2rem 0;
-}
-.title {
-  margin: 0;
-  font-size: $font-size-xxl;
-  font-weight: bold;
-  margin-bottom: 1rem;
-  @include ellipsisText(1, $font-size-xxl);
-}
-.basic {
-  margin: 0;
-  font-size: $font-size-sm;
-}
