+++ components/AlocholType/AlcoholTypesimilarList/AlcoholTypesimilarList.module.css
@@ -1,0 +1,62 @@
+.container{
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    padding: 0.5rem;
+    width: 100%;
+    height: auto;
+    margin: 1.5rem 0 2rem 0;
+}
+.title{
+    margin: 0;
+    font-size: 1.8rem;
+    font-weight: bold;
+    color: #000;
+    margin-bottom: 1rem;
+}
+.cardList{
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    flex-wrap: nowrap;
+    gap: 10px;
+    padding: 0.5rem;
+    width: auto;
+    height: auto;
+}
+.card{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 5%;
+    padding: 0.5rem;
+    width: 30%;
+    border-radius: 20px;
+}
+.card:hover{
+    box-shadow: 1px 1px 4px rgba(0,0,0,0.1);
+}
+.leftBox{
+    width: 100px;
+    aspect-ratio: 1;
+    height: fit-content;
+    border-radius: 10px;
+}
+.image{
+    width: 100%;
+    object-fit: cover;
+    aspect-ratio: 1;
+    border-radius: 10px;
+}
+.name{
+    margin: 0;
+    font-size: 1.5rem;
+    font-weight: bold;
+    color: #000;
+    margin-bottom: 1rem;
+}
+.basic{
+    margin: 0;
+    font-size: 1rem;
+}
