+++ components/AlocholType/SearchBar/SearchBar.module.css
@@ -1,0 +1,38 @@
+.searchBarBox{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    align-self: center;
+    padding: 5rem 0;
+    height: 50px;
+    border-radius: 20px;
+    text-align: center;
+}
+.searchBar{
+    display: flex;
+    align-items: center;
+    width: 80%;
+    box-shadow: rgba(0,0,0,0.3) 1px 1px 1px 1px;
+    border-radius: 20px;
+    border: rgba(0,0,0,0.3) 1px solid;
+}
+.searchInput{
+    outline: rgba(0,0,0,0.5) 1px solid;
+    background-color:white;
+    color: #000;
+    width: 90%;
+    height: 3rem;
+    border-radius: 20px 0 0 20px;
+
+    font-size: 1.2rem;
+    font-weight: regular;
+    text-align: left;
+    padding-left: 1.3rem;
+}
+.searchButton{
+    width: 10%;
+    height: 100%;
+    background-color: #000;
+    border-radius: 0 20px 20px 0;
+}
