--- styles/AlcoholType/ATPageStyle.module.scss
@@ -1,22 +1,0 @@
-@import '~styles/theme.scss';
-
-.cardListMain {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-gap: 10px;
-  width: 100%;
-  @media screen and (max-width: $breakpoint-sm) {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-  }
-}
-.cardListTitle {
-  font-size: $font-size-xl;
-  font-weight: bold;
-  padding: 1rem 0;
-  margin: 1rem 1rem 0.8rem 0rem;
-  &:first-of-type {
-    margin-top: 0;
-  }
-}
