+++ styles/AlcoholType/AlcoholType.module.css
@@ -1,0 +1,12 @@
+.cardListMain{
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    grid-gap: 10px;
+    width: 100%;
+}
+.cardListTitle{
+    font-size: 1.5rem;
+    font-weight: bold;
+    color: #222222;
+    padding: 1rem 0;
+}
