+++ styles/theme.scss
@@ -6,17 +6,15 @@ $black: #000; // font-color
 $dark-gray: #70757a; // font-color(for placeholer)
 $white: #fff; // background-color
 
-$gray11: #b7b7b7; // font-color(for placeholer)
-
 // Font Size ==========================
-$font-size-xxs: 0.75rem; // 12px
-$font-size-xs: 0.875rem; // 14px
-$font-size-sm: 1rem; // 16px
-$font-size-md: 1.125rem; // 18px
-$font-size-lg: 1.25rem; // 20px
-$font-size-xl: 1.5rem; // 24px
-$font-size-xxl: 1.75rem; // 28px
-$font-size-xxxl: 2rem; // 32px
+$font-size-xxs : 0.75rem; // 12px
+$font-size-xs : 0.875rem; // 14px
+$font-size-sm : 1rem; // 16px
+$font-size-md : 1.125rem; // 18px
+$font-size-lg : 1.25rem; // 20px
+$font-size-xl : 1.5rem; // 24px
+$font-size-xxl : 1.75rem; // 28px
+$font-size-xxxl : 2rem; // 32px
 
 // Font-Weight ========================
 $font-weight-bold: 700;
@@ -28,15 +26,3 @@ $font-weight-thin: 100;
 // responsive ========================
 $breakpoint-sm: 768px;
 $breakpoint-lg: 1060px;
-
-// Mixin =============================
-@mixin ellipsisText($line-cnt, $line-height) {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: $line-cnt; /* 라인수 */
-  -webkit-box-orient: vertical;
-  word-wrap: break-word;
-  line-height: $line-height;
-  height: $line-height * $line-cnt; /* line-height 가 1.2em 이고 3라인을 자르기 때문에 height는 1.2em * 3 = 3.6em */
-}
